[#M10] feat : 폴더 정리
