+++ prisma/mock.js
@@ -30,7 +30,6 @@ export const ARTICLES = [
     content: "판다인형 구매 후기입니다.",
     imageUrl: "https://sitem.ssgcdn.com/62/11/49/item/1000559491162_i1_1100.jpg",
     likeCount: 7,
-    isLiked: false,
     writer: "판다인형 수집가",
   },
   {
@@ -39,7 +38,6 @@ export const ARTICLES = [
     content: "판다인형 판매 후기입니다.",
     imageUrl: "https://sitem.ssgcdn.com/62/11/49/item/1000559491162_i1_1100.jpg",
     likeCount: 2,
-    isLiked: true,
     writer: "판다인형 중개인",
   },
   {
@@ -48,7 +46,6 @@ export const ARTICLES = [
     content: "불곰인형 구하는 곳 아시는분 계신가요?",
     imageUrl: "https://wimg.mk.co.kr/meet/2021/09/image_listtop_2021_854860_1630738087.jpg",
     likeCount: 3,
-    isLiked: false,
     writer: "불곰인형 수집가",
   },
 ];
